# swagger-spec.json
{
  "swagger": "2.0",
  "basePath": "/api",
  "paths": {
...
    "/avaliacoes": {
      "post": {
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/AvaliacaoDesastre"
            }
          }
        },
        "summary": "Criar uma nova avaliação de desastre",
        "operationId": "criar_avaliacao",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/EntradaAvaliacao"
            }
          },
          {
            "name": "X-Fields",
            "in": "header",
            "type": "string",
            "format": "mask",
            "description": "An optional fields mask"
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "avaliacoes"
        ]
      },
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/AvaliacaoDesastre"
              }
            }
          }
        },
        "summary": "Listar todas as avaliações de desastre",
        "operationId": "listar_avaliacoes",
        "parameters": [
          {
            "enum": [
              "agua_potavel",
              "alimentacao",
              "abrigo_temporario",
              "roupas_cobertores",
              "medicamentos",
              "outros"
            ],
            "in": "query",
            "description": "Filtrar por necessidade urgente",
            "name": "urgent_need",
            "type": "string"
          },
          {
            "enum": [
              "habitacao",
              "comercio",
              "agricultura",
              "outro"
            ],
            "in": "query",
            "description": "Filtrar por tipo de estrutura",
            "name": "structure_type",
            "type": "string"
          },
          {
            "enum": [
              "parcial",
              "grave",
              "total"
            ],
            "in": "query",
            "description": "Filtrar por nível de danos",
            "name": "damage_level",
            "type": "string"
          },
          {
            "type": "integer",
            "default": 10,
            "in": "query",
            "description": "Itens por página",
            "name": "per_page"
          },
          {
            "type": "integer",
            "default": 1,
            "in": "query",
            "description": "Número da página",
            "name": "page"
          },
          {
            "name": "X-Fields",
            "in": "header",
            "type": "string",
            "format": "mask",
            "description": "An optional fields mask"
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "avaliacoes"
        ]
      }
    },
...
  },
  "info": {
    "title": "API de Avaliação de Desastres",
    "version": "1.0",
    "description": "API para gestão de avaliações de desastres naturais"
  },
  "produces": [
    "application/json"
  ],
  "consumes": [
    "application/json"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Adicione: Bearer &lt;seu_token_jwt&gt;"
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
...
  "definitions": {
    "EntradaAvaliacao": {
      "required": [
        "contacto_telefonico",
        "endereco_completo",
        "membros_agregado",
        "necessidade_urgente",
        "nivel_danos",
        "nome_responsavel",
        "numero_documento",
        "tipo_estrutura"
      ],
      "properties": {
        "nome_responsavel": {
          "type": "string",
          "description": "Nome do Responsável"
        },
        "numero_documento": {
          "type": "string",
          "description": "Número de Documento (BI ou Passaporte)"
        },
        "contacto_telefonico": {
          "type": "string",
          "description": "Contacto Telefónico"
        },
        "membros_agregado": {
          "type": "integer",
          "description": "N.º de Pessoas no Agregado Familiar"
        },
        "grupos_vulneraveis": {
          "type": "array",
          "description": "Grupos Vulneráveis",
          "items": {
            "type": "string"
          }
        },
        "endereco_completo": {
          "type": "string",
          "description": "Endereço Completo"
        },
        "ponto_referencia": {
          "type": "string",
          "description": "Ponto de Referência"
        },
        "latitude_gps": {
          "type": "number",
          "description": "Latitude GPS"
        },
        "longitude_gps": {
          "type": "number",
          "description": "Longitude GPS"
        },
        "tipo_estrutura": {
          "type": "string",
          "description": "Tipo de Estrutura Afetada",
          "example": "habitacao",
          "enum": [
            "habitacao",
            "comercio",
            "agricultura",
            "outro"
          ]
        },
        "nivel_danos": {
          "type": "string",
          "description": "Nível de Danos",
          "example": "parcial",
          "enum": [
            "parcial",
            "grave",
            "total"
          ]
        },
        "perdas": {
          "type": "array",
          "description": "Perdas",
          "items": {
            "type": "string"
          }
        },
        "outras_perdas": {
          "type": "string",
          "description": "Especificação de Outras Perdas"
        },
        "necessidade_urgente": {
          "type": "string",
          "description": "Necessidade Urgente",
          "example": "agua_potavel",
          "enum": [
            "agua_potavel",
            "alimentacao",
            "abrigo_temporario",
            "roupas_cobertores",
            "medicamentos",
            "outros"
          ]
        },
        "outra_necessidade": {
          "type": "string",
          "description": "Especificação de Outra Necessidade Urgente"
        }
      },
      "type": "object"
    },
    "AvaliacaoDesastre": {
      "required": [
        "contacto_telefonico",
        "endereco_completo",
        "membros_agregado",
        "necessidade_urgente",
        "nivel_danos",
        "nome_responsavel",
        "numero_documento",
        "tipo_estrutura"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "description": "ID da Avaliação",
          "readOnly": true
        },
        "nome_responsavel": {
          "type": "string",
          "description": "Nome do Responsável"
        },
        "numero_documento": {
          "type": "string",
          "description": "Número de Documento (BI ou Passaporte)"
        },
        "contacto_telefonico": {
          "type": "string",
          "description": "Contacto Telefónico"
        },
        "membros_agregado": {
          "type": "integer",
          "description": "N.º de Pessoas no Agregado Familiar"
        },
        "grupos_vulneraveis": {
          "type": "array",
          "description": "Grupos Vulneráveis",
          "items": {
            "type": "string"
          }
        },
        "endereco_completo": {
          "type": "string",
          "description": "Endereço Completo"
        },
        "ponto_referencia": {
          "type": "string",
          "description": "Ponto de Referência"
        },
        "latitude_gps": {
          "type": "number",
          "description": "Latitude GPS"
        },
        "longitude_gps": {
          "type": "number",
          "description": "Longitude GPS"
        },
        "tipo_estrutura": {
          "type": "string",
          "description": "Tipo de Estrutura Afetada",
          "example": "habitacao",
          "enum": [
            "habitacao",
            "comercio",
            "agricultura",
            "outro"
          ]
        },
        "nivel_danos": {
          "type": "string",
          "description": "Nível de Danos",
          "example": "parcial",
          "enum": [
            "parcial",
            "grave",
            "total"
          ]
        },
        "perdas": {
          "type": "array",
          "description": "Perdas",
          "items": {
            "type": "string"
          }
        },
        "outras_perdas": {
          "type": "string",
          "description": "Especificação de Outras Perdas"
        },
        "ficheiros_prova": {
          "type": "array",
          "description": "Ficheiros de prova",
          "items": {
            "type": "string"
          }
        },
        "necessidade_urgente": {
          "type": "string",
          "description": "Necessidade Urgente",
          "example": "agua_potavel",
          "enum": [
            "agua_potavel",
            "alimentacao",
            "abrigo_temporario",
            "roupas_cobertores",
            "medicamentos",
            "outros"
          ]
        },
        "outra_necessidade": {
          "type": "string",
          "description": "Especificação de Outra Necessidade Urgente"
        },
        "data_criacao": {
          "type": "string",
          "format": "date-time",
          "description": "Data de Criação",
          "readOnly": true
        },
        "data_atualizacao": {
          "type": "string",
          "format": "date-time",
          "description": "Data da Última Atualização",
          "readOnly": true
        }
      },
      "type": "object"
    },
...
  },
...
}